+++ packages/celer-web-app/src/core/compiler/text/index.ts
@@ -1,2 +1,3 @@
 export * from "./TypedString";
 export {default as StringParser} from "./StringParser";
+export * from "./interpolation";